requirements.txt
html5lib
lxml
requests
python-dotenv
pandas
numpy
PyMySQL
mysql-connector
sqlalchemy
schedule